requirements.txt
# Wan2.2-Animate-Mix Video Face Swapper Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Cloud storage (optional, for OSS support)
oss2>=2.18.0

# Utilities
urllib3>=2.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
